requirements.txt
streamlit
google-generativeai
transformers
torch
pandas